Refactor MockServer; move to separate file
